requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
Flask==3.0.3
fonttools==4.53.1
frozendict==2.4.4
html5lib==1.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
multitasking==0.0.11
nltk==3.9.1
numpy==2.1.0
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
peewee==3.17.6
pillow==10.4.0
platformdirs==4.2.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==1.3.1
scipy==1.14.1
six==1.16.0
soupsieve==2.6
textblob==0.18.0.post0
tqdm==4.66.5
tweepy==4.14.0
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
Werkzeug==3.0.4
yfinance==0.2.43
